// src/Interfaces/InterfaceAuctionListing.tsx
export interface InterfaceAuctionListing {
    auctionId: number;
    title: string;
    description: string;
    endDate: string;
    categoryId: number;
    reserve: number;
    sellerId: number;
    sellerFirstName: string;
    sellerLastName: string;
    numBids: number;
    highestBid: number;
}